+++ CHANGELOG.md
@@ -29,6 +29,10 @@
 
 - Set default params for `EventDeletionRequest` and `Contact`
 
+### Added
+
+- Expose `NostrParser` (https://github.com/rust-nostr/nostr-sdk-ffi/pull/13)
+
 ## v0.42.1 - 2025/05/26
 
 ### Changed