+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Changed
 
 - Set default params for `EventDeletionRequest` and `Contact`
+- Update the android libraries to use 16KB page alignment (https://github.com/rust-nostr/nostr-sdk-ffi/pull/18)
 
 ## v0.42.1 - 2025/05/26